+++ arrays/shuffle/basic.js
@@ -3,10 +3,10 @@
  * @param {Array} arr
  */
 module.exports = arr => {
-    const lim = arr.length;
+    const lim = arr.length
     for (let i = 0; i < lim; i++) {
-        const j = Math.floor(Math.random() * lim);
-        [arr[i], arr[j]] = [arr[j], arr[i]];
+        const j = Math.floor(Math.random() * lim)
+        [arr[i], arr[j]] = [arr[j], arr[i]]
     }
-    return arr;
-};
+    return arr
+}