+++ maths/minSquaresToFill/basic.js
@@ -5,15 +5,15 @@
  * @returns {object}
  */
 module.exports.bigintCheck = (x, y) => {
-    x = BigInt(x), y = BigInt(y);
-    const res = { steps: [], x, y, final: 0n };
+    x = BigInt(x), y = BigInt(y)
+    const res = { steps: [], x, y, final: 0n }
     while (x && y) {
-        const min = x > y ? y : x;
-        const ratio = (x > y ? x : y) / min;
-        res.steps.push(ratio) && (res.final += ratio) && (x > y ? x -= ratio * min : y -= ratio * min);
+        const min = x > y ? y : x
+        const ratio = (x > y ? x : y) / min
+        res.steps.push(ratio) && (res.final += ratio) && (x > y ? x -= ratio * min : y -= ratio * min)
     }
-    return res;
-};
+    return res
+}
 
 /**
  * Gives the minimum amount of squares found to fill a rectangle of `x` and `y` dimensions.
@@ -22,14 +22,14 @@ module.exports.bigintCheck = (x, y) => {
  * @returns {object}
  */
 module.exports.intCheck = (x, y) => {
-    const res = { steps: [], x, y, final: 0 };
+    const res = { steps: [], x, y, final: 0 }
     while (x && y) {
-        const min = x > y ? y : x;
-        const ratio = Math.floor((x > y ? x : y) / min);
-        res.steps.push(ratio) && (res.final += ratio) && (x > y ? x -= ratio * min : y -= ratio * min);
+        const min = x > y ? y : x
+        const ratio = Math.floor((x > y ? x : y) / min)
+        res.steps.push(ratio) && (res.final += ratio) && (x > y ? x -= ratio * min : y -= ratio * min)
     }
-    return res;
-};
+    return res
+}
 
 /**
  * Gives the minimum amount of squares found to fill a rectangle of `x` and `y` dimensions.
@@ -37,4 +37,4 @@ module.exports.intCheck = (x, y) => {
  * @param {number|bigint} y
  * @returns {object}
  */
-module.exports = (x, y) => typeof x == "bigint" || typeof y == "bigint" ? this.bigintCheck(x, y) : this.intCheck(x, y);
+module.exports = (x, y) => typeof x == "bigint" || typeof y == "bigint" ? this.bigintCheck(x, y) : this.intCheck(x, y)