+++ objects/getObjValue/basic.js
@@ -5,14 +5,14 @@
  * @returns {object}
  */
 module.exports = (obj, ...keys) => {
-    if (keys.length == 1) keys = keys.flat();
-    if (typeof keys == "string") keys = keys.split(".");
+    if (keys.length == 1) keys = keys.flat()
+    if (typeof keys == "string") keys = keys.split(".")
 
-    const res = { current: obj, steps: [] };
+    const res = { current: obj, steps: [] }
     for (const k of keys) {
-        res.current = res.current[k];
-        if (res.current == null) break;
-        res.steps.push(res.current);
+        res.current = res.current[k]
+        if (res.current == null) break
+        res.steps.push(res.current)
     }
-    return res;
-};
+    return res
+}