+++ others/convertES/basic.js
@@ -8,5 +8,5 @@ module.exports = (str, goal = "ES6") => {
     return goal == "ES6" ? str.replace(
         /import\s+(?:(?:([A-Za-z_$][\w$]*)|\*\s+as\s+([A-Za-z_$][\w$]*)|({\s*[A-Za-z_$][\w$]*(?:\s+as\s+[A-Za-z_$][\w$]*)?\s*(?:,\s*[A-Za-z_$][\w$]*(?:\s+as\s+[A-Za-z_$][\w$]*)?\s*)?}))\s+from\s+(("|'|`).*\5)|(("|'|`).*\7))/g,
         (_, ...a) => a[0] || a[1] || a[2] ? `const ${a[0] || a[1] || a[2]} = require(${a[3]})` : a[5] && `require(${a[5]})`
-    ).replace(/([A-Za-z_$][\w$]*)\s+as\s+([A-Za-z_$][\w$]*)/g, "$1: $2") : null;
-};
+    ).replace(/([A-Za-z_$][\w$]*)\s+as\s+([A-Za-z_$][\w$]*)/g, "$1: $2") : null
+}