+++ others/redo/minified.js
@@ -1,0 +1,1 @@
+function redo(thisArg,fn,cond){if(cond(thisArg)){thisArg=fn(thisArg);return redo(thisArg,fn,cond)}return thisArg}module.exports = redo