+++ strings/args/basic.js
@@ -4,6 +4,6 @@
  * @param {number} group Maximum groups' length. 0 for Infinity (same as `<String>.split()`). Default is 1.
  */
 module.exports = (str, group = 1) => {
-    if (isNaN(group) || group < 0) return null;
-    return str.match(RegExp(`\\S+(?:\\s+\\S+){${group - 1}}|\\S+(\\s+\\S+)*`, "g"));
-};
+    if (isNaN(group) || group < 0) return null
+    return str.match(RegExp(`\\S+(?:\\s+\\S+){${group - 1}}|\\S+(\\s+\\S+)*`, "g"))
+}